tests: Rename argon2i testing file
